Bind HelpMenu to a button in the QDialogButtonBox
Not sure if it's a good thing to kill menu bars in general. But in
this case it makes sense in my opinion because the help menu was the
only menu in the menu bar. Furthermore the button box already
provided the space in the UI and thus we gain a few pixels in
vertical direction.