Don't enable plugins with no chance to disable them
Admittedly I don't know the motivation behind the current situation,
but if an UI is provided to enable/disable plugins it seems a bit
inconsisten to hard enable some plugins with no possibility for the
user to disable them. This might even be confusing because users
might take the disabled checkbox (even though it's checked) as an
indication that the plugin can't be enabled.
To avoid this, have plugins enabled by default instead of always
enabled.

BUG: 396569