fix dangerously implicit int->int->enum mapping
the type order in the qstringlist was implicitly the int values of the
type enum. incredibly scary. instead push the items into the combobox
along with their enum value as Qt::UserRole and then pass that along.
this ensures (kind of) that the int we cast back to enum in
initKRenameVars will be an actual enum value