port from deprecated kjs to qjsengine
for kf6 todo https://phabricator.kde.org/T11604

kjs is deprecated and we want to move away from it. this ports to
qjsengine which does what we need for krename just fine and is also
supported in qt6 still

all use cases of the js plugin still seem to work fine. variable
definition, function definition and using either in the replacement.
same for the implicit global variables.