Fix a capitalisation during krename --test
dff7935ed121c183d2c1a13c2f0696fad15a638d made all occurrence of krename
adhere to the same style. In a test (invoked via krename --test)
it isn't used as the application name but as the name of a directory
to test some tokens, leading to a failing test.