cmake: Improve FindPoDoFo
Add some rst based documentation and provide an imported target for
example.
I've dropped -DUSING_SHARED_PODOFO, which looks weird nowadays and
seems to exist because of Windows, which I can't test. So comment it out
for now, but it'll probably be removed in the future, if nobody turns up
who cares about Windows and can comment on it.