Support podofo 0.10
Version 0.10 of podofo is a complete rewrite. krename's use of it is minimal, so porting is easy.

Switch the cmake module to use pkgconfig, which is available since 0.9.5 (release in 2017).

Unfortunately, the hack to find the version number is still needed, since the pc file is buggy and ships an empty "Version" field.